apiclient/requirements.txt
mwrogue
python-dotenv
